UT: Extract SingleFileRepository from test_pba_file_upload.py
