+++ monkey/tests/monkey_island/__init__.py
@@ -1,0 +1,1 @@
+from .single_file_repository import SingleFileRepository
